+++ src/core/internal/delegation.ts
@@ -18,7 +18,7 @@ import {
   execute as execute_viem,
 } from 'viem/experimental/erc7821'
 
-import type * as DelegatedAccount from './account.js'
+import * as DelegatedAccount from './account.js'
 import { delegationAbi } from './generated.js'
 import * as Key from './key.js'
 import type { OneOf } from './types.js'
@@ -57,8 +57,7 @@ export async function execute<
       client,
       parameters,
     )
-    const signatures = await sign({
-      account,
+    const signatures = await DelegatedAccount.sign(account, {
       key,
       payloads,
     })
@@ -433,86 +432,4 @@ export declare namespace keyAt {
      */
     index: number
   }
-}
-
-/**
- * Extracts a signing key from a delegated account and signs payload(s).
- *
- * @example
- * TODO
- *
- * @param parameters - Parameters.
- * @returns Signatures.
- */
-export async function sign(
-  parameters: sign.Parameters,
-): Promise<sign.ReturnType> {
-  const { account, key, payloads } = parameters
-
-  const [payload, authorizationPayload] = payloads
-
-  // In order to sign (and perform) an authorization, we need the EOA's root key.
-  // We will extract an "owner" key from either the `key` parameter or the provided `account`.
-  const ownerKey = (() => {
-    // Extract from `key` parameter.
-    if (typeof key === 'object' && key.role === 'owner') return key
-    if (typeof key === 'number' && account.keys?.[key]?.role === 'owner')
-      return account.keys[key]
-
-    // Extract from the `account`.
-    return account.keys?.find((key) => key.role === 'owner')
-  })()
-
-  // If we have an authorization payload, but no "owner" key on the account,
-  // then we cannot perform an authorization as we need the EOA's private key.
-  if (authorizationPayload && !ownerKey?.sign)
-    throw new Error('account does not have an owner key.')
-
-  // Extract a key to sign the payload with.
-  const signingKey = (() => {
-    // Extract from `key` parameter.
-    if (typeof key === 'object') return key
-
-    // Extract from the `account` (with optional `key` index).
-    if (!account.keys) return undefined
-    if (typeof key === 'number') return account.keys[key]
-    return account.keys[0]
-  })()
-
-  // If the account has no valid signing key, then we cannot sign the payload.
-  if (!signingKey || !signingKey.sign)
-    throw new Error('cannot find key to sign with.')
-
-  // Sign the payload(s).
-  const signatures = await Promise.all([
-    signingKey.sign({ payload }),
-    authorizationPayload && ownerKey?.sign
-      ? ownerKey.sign({ payload: authorizationPayload })
-      : undefined,
-  ])
-
-  return signatures as never
-}
-
-export declare namespace sign {
-  type Parameters = {
-    /**
-     * The delegated account to sign the payloads for.
-     */
-    account: DelegatedAccount.Account
-    /**
-     * Key to sign the payloads with. If not provided, a key will be extracted from the `account`.
-     */
-    key?: number | Key.Key | undefined
-    /**
-     * Payloads to sign.
-     */
-    payloads:
-      | [executePayload: Hex.Hex]
-      | [executePayload: Hex.Hex, authorizationPayload: Hex.Hex]
-  }
-
-  type ReturnType =
-    | [executeSignature: Hex.Hex]
-    | [executeSignature: Hex.Hex, authorizationSignature: Hex.Hex]
 }
